+++ src/apps/front/src/features/feed/_feed.scss
@@ -36,11 +36,11 @@ main {
   text-align: center;
 }
 
-.nav-main > div {
+.nav-main>div {
   margin-bottom: 20px;
 }
 
-.nav-battom > .setting {
+.nav-battom>.setting {
   margin-bottom: 7px;
 }
 
@@ -53,11 +53,11 @@ main {
   height: 24px;
 }
 
-.nav-main > .white_hover {
+.nav-main>.white_hover {
   padding: 13px;
 }
 
-.nav-battom > .white_hover {
+.nav-battom>.white_hover {
   padding: 13px;
 }
 
@@ -83,6 +83,9 @@ main {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
+  position: fixed;
+  top: 0;
+  left: 0;
 }
 
 .imagination {
@@ -101,7 +104,7 @@ main {
   box-shadow: 0px 0px 10px 4px rgba(0, 0, 0, 0.1);
 }
 
-.fropile > .Rectangle {
+.fropile>.Rectangle {
   width: 125px;
   height: 127px;
 }
@@ -118,11 +121,11 @@ main {
   border-right: 1px solid #d9d9d9;
 }
 
-.fropile-cnt > div > p:nth-child(2) {
+.fropile-cnt>div>p:nth-child(2) {
   font-size: 10px;
 }
 
-.fropile-cnt > div {
+.fropile-cnt>div {
   padding: 5px 10px;
 }
 
@@ -140,7 +143,7 @@ main {
   padding: 20px 0px;
 }
 
-.open-nav-icon > .hoverwhite {
+.open-nav-icon>.hoverwhite {
   margin-bottom: 15px;
 }
 
@@ -158,11 +161,11 @@ main {
   box-shadow: 0px 0px 10px 4px rgba(0, 0, 0, 0.1);
 }
 
-.hoverwhite > .nav-icon {
+.hoverwhite>.nav-icon {
   margin-right: 20px;
 }
 
-.fropile-cnt > div > p:nth-child(2) {
+.fropile-cnt>div>p:nth-child(2) {
   color: #b9b9b9;
   font-family: Pretendard;
   font-size: 10px;
@@ -179,6 +182,61 @@ main {
   padding-right: 22px;
 }
 
-.open-nav-battom > .setting {
+.open-nav-battom>.setting {
   margin-bottom: 2px;
+}
+
+
+// Lnb css
+.Lnb {
+  margin-top: 20px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 85%;
+}
+
+.create-btn {
+  padding: 11px 20px;
+  border-radius: 10px;
+  border: 1px solid #E0E0E0;
+  background: #0288D1;
+  color: white;
+}
+
+.search-icon {
+  position: relative;
+  right: 35px;
+  top: 5px;
+}
+
+.form {
+  display: inline-block;
+}
+
+.search-input {
+  width: 589px;
+  height: 39px;
+  flex-shrink: 0;
+  border-radius: 10px;
+  background: #FFF;
+  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.25);
+  background-image: url('./img/search.png');
+  background-position: calc(100% - 15px) center;
+  background-repeat: no-repeat;
+}
+.chat{
+  width: 30px;
+  height: 30px;
+  margin-left: 20px;
+}
+.alarm{
+  width: 16px;
+  height: 19.5px;
+  margin-bottom: 5px;
+}
+.layout-position{
+  position: relative;
+  bottom: 60px;
+  right: 200px;
 }
