+++ src/apps/front/src/features/feed/components/Lnb.tsx
@@ -3,10 +3,17 @@ import { Link } from 'react-router-dom';
 
 const Lnb = () => {
   return (
-    <div>
-      <Link to={'/create'}>create</Link>
-      <button>알림창</button>
-      <Link to={'/chat'}>chat</Link>
+    <div className='Lnb'>
+      <form action="" className='form'>
+        <input type="text" className='search-input' />
+      </form>
+      <button className='create-btn'>
+        <Link to={'/create'}>+ Create</Link>
+      </button>
+      <div className='allim'>
+        <img src="./img/alarm.png" alt="" className='alarm' />
+        <img src="./img/share.png" alt="" className='chat'/>
+      </div>
     </div>
   );
 };
